rename MessageForm + organize some components
