+++ src/components/layout/BasicTemplate.vue
@@ -1,13 +1,9 @@
 <template>
-  <div class="submit-question">
-    <div class="basic-template">
-      <content-header :header-title="headerTitle" />
-      <div class="basic-template__content">
-        <div class="content-card">
-          <div class="content-card__card">
-            <slot />
-          </div>
-        </div>
+  <div class="basic-template">
+    <content-header :header-title="headerTitle" />
+    <div class="content-card">
+      <div class="content-card__card">
+        <slot />
       </div>
     </div>
   </div>
@@ -37,22 +33,17 @@ export default {
 }
 </script>
 
-<style lang="scss">
-.submit-question {
-  height: 100vh;
-}
-
+<style lang="scss" scoped>
 .basic-template {
   min-height: 100vh;
   background: $c-backdrop;
-}
-.basic-template__content {
 }
 
 .content-card {
   padding: 80px 0;
   width: 100%;
 }
+
 .content-card__card {
   background: #fff;
   width: 80%;
